requirements.txt
fer==22.4.0
numpy==1.23.2
opencv_python==4.6.0.66
Pillow==9.2.0
requests==2.28.1
selenium==4.4.3
webdriver_manager==3.8.3
